# Improve performance for DefaultExecutionPlan.hasOverlap (fix #11792)