Fixes #11792

### Context
Improves the performance of DefaultExecutionPlan.hasOverlap by an order of magnitude when there are many items in both sets. For small sets, it got a bit slower in my tests, but this should be negligible I hope.

Detailed timings from my micro-benchmarks:
- 10 items x 10 items: before: 0.0004ms, after: 0.0009ms
- 10000 items x 10 items: before: 0.9ms, after: 0.9ms
- 10 items x 10000 items: before: 1ms, after: 1.6ms
- 10000 items x 10000 items: before: 604ms, after: 4ms

The timings even for the large sets are not as horrible as they were in #11792, so either the sets were even larger in my build, or the performance got worse due to long paths (I used rather short strings in my tests).

### Contributor Checklist
- [X] [Review Contribution Guidelines](https://github.com/gradle/gradle/blob/master/CONTRIBUTING.md)
- [X] Make sure that all commits are [signed off](https://git-scm.com/docs/git-commit#git-commit---signoff) to indicate that you agree to the terms of [Developer Certificate of Origin](https://developercertificate.org/).
- [X] Check ["Allow edit from maintainers" option](https://help.github.com/articles/allowing-changes-to-a-pull-request-branch-created-from-a-fork/) in pull request so that additional changes can be pushed by Gradle team
- [N/A] Provide integration tests (under `<subproject>/src/integTest`) to verify changes from a user perspective
- [X] Provide unit tests (under `<subproject>/src/test`) to verify logic
- [N/A] Update User Guide, DSL Reference, and Javadoc for public-facing changes
- [X] Ensure that tests pass locally: `./gradlew <changed-subproject>:check`

### Gradle Core Team Checklist
- [ ] Verify design and implementation 
- [ ] Verify test coverage and CI build status
- [ ] Verify documentation
- [ ] Recognize contributor in release notes