@foto-andreas 2020-05-06
Hi,

still not included in Rev 6.4? Is something missing to merge the PR?

Andreas