@tobiasbaum 2020-05-11
This PR is now likely obsolete, due to the alternative fix in PR #13066